Make key store types extensible conveniently
